move one time key errors out of model
